WIP: interpolate radar sampling time
This was something I always wanted to include when gridding radar data,
as time information is vital for advanced algorithms (e.g., tracking). I
have implemented this time interpolation, and have done some analysis of
the results, but I am still not 100% sure this is implemented properly,
so more effort needs to be put into this. It is a good start though...