WIP: adding grid quality index
Adding a new potential output when gridding radar data that is designed
to help with identifying potentially poorly gridded (constrained)
regions. The essential input is a mask for a particular radar field
which indicates where there are valid gates (a value of True or 1) and
where where there are invalid (missing) gates (a value of False or 0).
The idea here is that a grid point in which we have "full" confidence in
will have a grid quality index that approaches 1, and a grid point in
which we have "no" confidence in will have a grid quality index that
approaches 0.