WIP: significant improvements to mapper
Significant changes and additions to the Weight class and radar mapper.
I have moved the Weight class to the new core module since this is a
core functionality of this package. Furthermore, the Weight class now
incorporates information not only about objective analysis parameters
but also about a radar and its kd-tree. These new additions were driven
primarily by the realization that a kd-tree for a particular radar
volume and its subsequent querying can be re-used for a similar radar
volume, greatly increasing processing time for scripts that are gridding
multiple similar radar volumes.

This commit also includes multiple additions to documentation of various
modules, but this is still an ongoing process, and I still should learn
Sphinx in order to improve this documentation.