ENH: Domain class
Adding an enhancement to the core module: the Domain class. This class
is designed to store all data pertinent to the rectilinear grid in which
the user wants to map his/her data onto. Thinking slightly ahead with
this class, it includes the 'dem' attribute which would contain a
gdal.Dataset if available. With this data we could include information
about the surface terrain when mapping radar data.

I have also added the common submodule to the interpolation module which
helps clean up the mapper submodule a bit, like moving the _populate
functions out of mapper and into common.

Finally, I have tried to clean up the documentation a bit, but still
lots more work to do here...