WIP: utility module
This is still a work in progress. I have moved util.standard_refraction
to util.equivalent_earth_model to be more consistent with the literature
and the underlying methodology. This function now includes the ability
to specify the Earth's radius (R) and the effective Earth radius multiplier
(ke). The default corresponds to standard atmospheric refraction as defined
in Doviak and Zrnic (1993).

More work is needed here to be consistent with map projection
information, as well as computing radar pointing directions on the
defined grid, i.e., Issue #2.